Revise Add & Update UI Components for Employee
